refactor: extract useSessionToast to @/hooks/ui/useSessionToast
Reusable hook for post-navigation toast feedback (org switch, invitation
acceptance, etc). Exports SESSION_TOAST_KEY for writers.
